% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{ex1pop}
\alias{ex1samp}
\title{Synthetic grouped two dimensional numeric (binned) data}
\format{
\code{grouped2d_pop} is a data frame with 100 cases (rows) and 2 variables (columns) named \code{Group} and
\code{X1}.
and \code{X2}.

\code{grouped2d_sample} is a data frame with 10 cases (rows) and 2 variables (columns) named \code{X1b} and
\code{X2b}.
}
\usage{
ex1pop

ex1samp
}
\description{
A synthetically created two dimensional
}
\keyword{datasets}
