% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmnchoose.R
\name{lmnchoose}
\alias{lmnchoose}
\title{Calculate the log multinomial coefficient.}
\usage{
lmnchoose(n, k)
}
\arguments{
\item{n}{number representing the size of the sample.}

\item{k}{numeric vector representing the number of times each category was chosen.}
}
\description{
\loadmathjax
\code{lmnchoose} returns the natural logarithm of the multinomial coefficient.
}
\details{
The multinomail coefficient is the
generalization of the binomial coefficient, and represents the number of ways it is possible to choose
\mjeqn{k_1}{k_1} objects of class 1, \mjeqn{k_2}{k_2} objects of class 2, up to \mjeqn{k_m}{k_m} objects of class m,
from a sample of size \code{n}.

The multinomail coefficient can be expressed as
\mjdeqn{{n \choose \mathbb{k}} = \frac{n!}{k_1! k_2! ... k_m!}}{}

This calculation is difficult (or impossible) for large \code{n}, so it is calculated in log space, which is much
easier. If the logarithm is not desired, the result is exponentiated, but beware this can also fail if \code{n} is
large enough.
}
