% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{loghastings_binomial_}
\alias{loghastings_binomial_}
\title{Calculate the log hastings adjustment for metropolis algorithm (quickly)}
\usage{
loghastings_binomial_(Ycur, Yprop, k)
}
\arguments{
\item{Ycur}{numeric vector for current value of population}

\item{Yprop}{numeric vector for proposed value of population}

\item{k}{position of the changing value}
}
\description{
Calculate the log hastings adjustment for metropolis algorithm (quickly)
}
