% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{loglik_binary_}
\alias{loglik_binary_}
\title{Calculate the log-likelihood for a 1-d binary population}
\usage{
loglik_binary_(N, n, Y, y)
}
\arguments{
\item{N}{the population size}

\item{n}{the sample size}

\item{Y}{the number of successes in the population}

\item{y}{the number of successes in the sample}
}
\description{
Calculate the log-likelihood for a 1-d binary population
}
