% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{loglik_binary_fast_}
\alias{loglik_binary_fast_}
\title{Calculate the log-likelihood for a 1-d binary population (quickly)}
\usage{
loglik_binary_fast_(P, n_0, n_1)
}
\arguments{
\item{P}{the population proportion}

\item{n_0}{the number of failures in the sample}

\item{n_1}{the number of successes in the sample}
}
\description{
Calculate the log-likelihood for a 1-d binary population (quickly)
}
