% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{loglik_multiclass_}
\alias{loglik_multiclass_}
\title{Calculate the log-likelihood for a 1-d multi-class population}
\usage{
loglik_multiclass_(n, Y, y)
}
\arguments{
\item{n}{the sample size.}

\item{Y}{a vector indicating the amount of each class in the population.}

\item{y}{a vector indicating the amount of each class in the sample.}
}
\description{
Calculate the log-likelihood for a 1-d multi-class population
}
