% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logpost_binary}
\alias{logpost_binary}
\title{Calculate the log-posterior of a 1-d binary population}
\usage{
logpost_binary(pop, obs, lprior)
}
\arguments{
\item{pop}{numeric vector representing the population}

\item{obs}{numeric vector representing the observed sample}

\item{lprior}{function uses to calculate the log-prior}
}
\description{
Calculate the log-posterior of a 1-d binary population
}
