% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logpost_binary_}
\alias{logpost_binary_}
\title{Calculate the log-posterior for a 1-d binary population}
\usage{
logpost_binary_(N, n, Y, y, lprior)
}
\arguments{
\item{N}{the population size}

\item{n}{the sample size}

\item{Y}{the number of successes in the population}

\item{y}{the number of successes in the sample}

\item{lprior}{function uses to calculate the log-prior}
}
\description{
Calculate the log-posterior for a 1-d binary population
}
