% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{logpost_binary_fast}
\alias{logpost_binary_fast}
\title{Calculate the log-posterior of a 1-d binary population (quickly)}
\usage{
logpost_binary_fast(pop, obs, lprior, a, b, ...)
}
\arguments{
\item{pop}{numeric vector representing the population}

\item{obs}{numeric vector representing the observed sample}

\item{lprior}{function uses to calculate the log-prior}

\item{a}{fix later}

\item{b}{fix later}

\item{...}{other arguments, passed to the lprior function}
}
\description{
Calculate the log-posterior of a 1-d binary population (quickly)
}
