% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{logpost_binary_fast_}
\alias{logpost_binary_fast_}
\title{Calculate the log-posterior for a 1-d binary population (quickly)}
\usage{
logpost_binary_fast_(P, n_0, n_1, lprior, ...)
}
\arguments{
\item{P}{the population proportion}

\item{n_0}{the number of failures in the sample}

\item{n_1}{the number of successes in the sample}

\item{lprior}{function uses to calculate the log-prior}

\item{...}{other arguments, passed to the lprior function}
}
\description{
Calculate the log-posterior for a 1-d binary population (quickly)
}
