% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logpost_multiclass}
\alias{logpost_multiclass}
\title{Calculate the log-posterior of a 1-d multi-class population}
\usage{
logpost_multiclass(pop, obs, lprior)
}
\arguments{
\item{pop}{numeric vector representing the population}

\item{obs}{numeric vector representing the observed sample}

\item{lprior}{function uses to calculate the log-prior}
}
\description{
Calculate the log-posterior of a 1-d multi-class population
}
