% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logpost_multiclass_}
\alias{logpost_multiclass_}
\title{Calculate the log-posterior for a 1-d multi-class population}
\usage{
logpost_multiclass_(N, n, Y, y, lprior)
}
\arguments{
\item{N}{the population size}

\item{n}{the sample size}

\item{Y}{a vector indicating the amount of each class in the population.}

\item{y}{a vector indicating the amount of each class in the sample.}

\item{lprior}{function uses to calculate the log-prior}
}
\description{
Calculate the log-posterior for a 1-d multi-class population
}
