% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{logprior_binary_beta_fast_}
\alias{logprior_binary_beta_fast_}
\title{Calculates the beta log-prior for a 1-d binary population (quickly)}
\usage{
logprior_binary_beta_fast_(P, a, b, ...)
}
\arguments{
\item{P}{the population proportion}

\item{a}{first shape parameter for beta prior}

\item{b}{second shape parameter for beta prior}

\item{...}{other arguments (unused)}
}
\description{
Calculates the beta log-prior for a 1-d binary population (quickly)
}
