% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logprior_binary_uninformative_}
\alias{logprior_binary_uninformative_}
\title{Calculates the uninformative log-prior for a 1-d binary population}
\usage{
logprior_binary_uninformative_(N, Y, ...)
}
\arguments{
\item{N}{the population size}

\item{Y}{the number of successes in the population}

\item{...}{other arguments (unused)}
}
\description{
Calculates the uninformative log-prior for a 1-d binary population
}
