% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logprior_multiclass_flat_}
\alias{logprior_multiclass_flat_}
\title{Calculates the flat log-prior for a 1-d multi-class population}
\usage{
logprior_multiclass_flat_(...)
}
\arguments{
\item{...}{other arguments (unused)}
}
\description{
Calculates the flat log-prior for a 1-d multi-class population
}
