% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_fast.R
\name{mcmc_binary_fast}
\alias{mcmc_binary_fast}
\title{Sample from the posterior of a 1-d binary population (quickly)}
\usage{
mcmc_binary_fast(obs, N, prior = "beta", control = list(...), ...)
}
\arguments{
\item{obs}{numeric vector representing the observed sample.}

\item{N}{size of the population.}

\item{prior}{choice of prior.}

\item{control}{a list of parameters for tuning the mcmc algorithm. Passed to \code{\link{lemur.control}}.}

\item{...}{arguments to be used to form the \code{control} object if it is not supplied directly.}
}
\description{
Sample from the posterior of a 1-d binary population (quickly)
}
