% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_superfast.R
\name{mcmc_binary_superfast}
\alias{mcmc_binary_superfast}
\title{Sample from the posterior predictive of a 1-d binary population (very quickly)}
\usage{
mcmc_binary_superfast(obs, N, nsamples, a, b)
}
\arguments{
\item{obs}{numeric vector representing the observed sample.}

\item{N}{size of the population.}

\item{nsamples}{number of samples to draw}

\item{a}{first hyperparameter for beta prior}

\item{b}{second hyperparameter for beta prior}
}
\description{
Sample from the posterior predictive of a 1-d binary population (very quickly)
}
