% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d_superfast.R
\name{mcmc_multinomial_superfast}
\alias{mcmc_multinomial_superfast}
\title{Sample from the posterior predictive of a 1-d multinomial population (very quickly)}
\usage{
mcmc_multinomial_superfast(obs, N, nsamples, alpha)
}
\arguments{
\item{obs}{numeric vector representing the observed sample.}

\item{N}{size of the population.}

\item{nsamples}{number of samples to draw}

\item{alpha}{hyperparameter for dirichlet prior; can be either a single number or a numeric vector}
}
\description{
Sample from the posterior predictive of a 1-d multinomial population (very quickly)
}
