% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped2d_data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{grouped2d_pop}
\alias{grouped2d_sample}
\title{Synthetic grouped two dimensional numeric data}
\format{
\code{grouped2d_pop} is a data frame with 5000 cases (rows) and 3 variables (columns) named \code{Group}, \code{X1},
and \code{X2}.

\code{grouped2d_sample} is a data frame with 100 cases (rows) and 3 variables (columns) named \code{Group},
\code{X1}, and \code{X2}.
}
\usage{
grouped2d_pop

grouped2d_sample
}
\description{
A synthetically created two dimensional numeric dataset containing 5000 observations (in the population) and 100
observations (in the sample), as well as a grouping variable with three levels. The \emph{true} correlations between the
two numeric variables are 0.6, 0.5, and 0.4 for group A, B, and C, respectively.
}
\keyword{datasets}
