% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim1d.R
\name{logprior_multiclass_uninformative_}
\alias{logprior_multiclass_uninformative_}
\title{Calculates the uninformative log-prior for a 1-d multi-class population}
\usage{
logprior_multiclass_uninformative_(N, Y)
}
\arguments{
\item{N}{the population size}

\item{Y}{a vector indicating the amount of each class in the population.}
}
\description{
Calculates the uninformative log-prior for a 1-d multi-class population
}
